Interrupt aware XPT2046_Touchscreen(uint8_t cs, uint8_t tirq)
On initializing with interrupt pin Poll calls for touch are ignored
unless preceded by falling edge on touch interrupt pin.

User calls are honored until touch detection records a failed pressure
test reading.

This allows the user code to call at a faster rate with no SPI impact so
that the start of a touch is not missed like when polled at a
conservative 100ms.

Included TouchTest sample modified for continuous loop() read to
demonstrate responsiveness and no SPI chatter, versus bright LED when
int pin is not used.